# spec/openapi.json
{
  "components": {
    "schemas": {
...
      "NoteRequest": {
        "description": "Request payload for creating or updating a note.",
        "properties": {
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "expectedRevision": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "noteId": {
            "type": [
              "string",
              "null"
            ]
          },
          "poiId": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "ProgressRequest": {
        "description": "Request payload for updating route progress.",
        "properties": {
          "expectedRevision": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "visitedStopIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 1000,
            "type": [
              "array",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "RouteAnnotationsResponse": {
        "description": "Response payload aggregating notes and progress for a route.",
        "properties": {
          "notes": {
            "items": {
              "$ref": "#/components/schemas/RouteNoteResponse"
            },
            "maxItems": 1000,
            "type": "array"
          },
          "progress": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RouteProgressResponse"
              }
            ]
          },
          "routeId": {
            "type": "string"
          }
        },
        "required": [
          "routeId",
          "notes"
        ],
        "type": "object"
      },
      "RouteNoteResponse": {
        "description": "Response payload for a route note.",
        "properties": {
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "poiId": {
            "type": [
              "string",
              "null"
            ]
          },
          "revision": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "routeId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "routeId",
          "body",
          "createdAt",
          "updatedAt",
          "revision"
        ],
        "type": "object"
      },
      "RouteProgressResponse": {
        "description": "Response payload for route progress.",
        "properties": {
          "revision": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "routeId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "visitedStopIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 1000,
            "type": "array"
          }
        },
        "required": [
          "routeId",
          "visitedStopIds",
          "updatedAt",
          "revision"
        ],
        "type": "object"
      },
...
      "crate.domain.Error": {
        "description": "OpenAPI schema for [`crate::domain::Error`].\n\nAPI error response payload with machine-readable code and human-readable\nmessage.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/crate.domain.ErrorCode",
            "description": "Stable machine-readable error code."
          },
          "details": {
            "description": "Supplementary error details for clients."
          },
          "message": {
            "description": "Human-readable message returned to clients.",
            "example": "Something went wrong",
            "type": "string"
          },
          "trace_id": {
            "description": "Correlation identifier for tracing this error across systems.",
            "example": "01HZY8B2W6X5Y7Z9ABCD1234",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "crate.domain.ErrorCode": {
        "description": "OpenAPI schema for [`crate::domain::ErrorCode`].\n\nStable machine-readable error codes returned in API error responses.",
        "enum": [
          "invalid_request",
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict",
          "service_unavailable",
          "internal_error"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Session cookie issued by POST /api/v1/login.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "HTTP interface for session-authenticated access and health probes.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "Wildside backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/routes/{route_id}/annotations": {
      "get": {
        "operationId": "getRouteAnnotations",
        "parameters": [
          {
            "description": "Route identifier",
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteAnnotationsResponse"
                }
              }
            },
            "description": "Route annotations"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Unauthorised"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Fetch notes and progress for a route.",
        "tags": [
          "routes"
        ]
      }
    },
    "/api/v1/routes/{route_id}/notes": {
      "post": {
        "operationId": "upsertRouteNote",
        "parameters": [
          {
            "description": "Route identifier",
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "UUID for idempotent requests",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteNoteResponse"
                }
              }
            },
            "description": "Upserted note"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Unauthorised"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Create or update a note for the route.",
        "tags": [
          "routes"
        ]
      }
    },
    "/api/v1/routes/{route_id}/progress": {
      "put": {
        "operationId": "updateRouteProgress",
        "parameters": [
          {
            "description": "Route identifier",
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "UUID for idempotent requests",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteProgressResponse"
                }
              }
            },
            "description": "Updated progress"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Unauthorised"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Update route progress for the authenticated user.",
        "tags": [
          "routes"
        ]
      }
    },
...
  },
  "security": [
    {
      "SessionCookie": []
    }
  ],
  "servers": [
    {
      "description": "Relative to the deployment base URL",
      "url": "/"
    }
  ],
...
}